+++ src/index.html
@@ -391,7 +391,6 @@
             <script src="./data/worldbank/worldbank.js"></script>
 		    <script src="js/data.js"></script>
 			<script src="js/main.js"></script>
-			<script src="js/ selectCreationFunctions.js"></script>
 			<!--<script src="main.js"></script>-->
     </body>
