+++ test/data.spec.js
@@ -1,49 +1,8 @@
 global.window = global;
 global.assert = require('chai').assert;
-require('../src/data');
+require('../src/js/data');
 require('./data.spec.js');
 
 
 
 
-// describe('aplicar', () => {
-
-//   const WORLDBANK = {
-//     "PER": {
-//       "indicators": [
-//         {
-//           "data":
-//           {
-//             "2013": 24.0900001525879,
-//             "2014": 24.2099990844727,
-//             "2015": 21.6800003051758,
-//             "2016": 23.7399997711182,
-//             "2017": 23.7600002288818
-//           },
-//           "countryName": "Perú",
-//           "countryCode": "PER",
-//           "indicatorName": "Empleo de tiempo parcial, mujeres (% del total de mujeres empleadas)"
-//         },
-//         "indicatorCode": "SL.TLF.PART.FE.ZS"}]
-//      }
-//   } 
-
-
-
-
-
-
-
-
-//   it('debería ser una función', () => {
-//     assert.equal(typeof aplicar, 'function');
-//   });
-
-//   it('debería retornar "aplicar"', () => {
-//     assert.deepEqual(aplicar(year, question, country), dataGral, ("2013", "Mujeres que creen que está justificado que un marido golpee a su esposa (cualquiera de las cinco razones) (%)", "PERU")[
-      
-      
-//   });
-
-
-// });
